+++ src/renderer/src/env.d.ts
@@ -1,0 +1,14 @@
+declare global {
+  interface Window {
+    lanpm: {
+      platform: NodeJS.Platform
+      versions: {
+        node: string
+        chrome: string
+        electron: string
+      }
+    }
+  }
+}
+
+export {}